feat(worker): Introduce skill processing with FlatBuffers serialization
- Added a new `capnproto-wrapper` subproject to handle Cap'n Proto dependencies.
- Updated `meson.build` files to include the new Cap'n Proto subproject.
- Implemented a skills processing module in the worker, allowing for typed task handlers.
- Created multiple skill handlers for various operations (e.g., string reversal, math operations, vector math, fused multiply-add).
- Refactored `TaskProcessor` and `SkillProcessor` to manage task and skill processing respectively.
- Updated runtime implementations to accommodate new skill-based processing.
- Enhanced error handling and logging for task processing.
- Added FlatBuffers schema generation for skill tasks.

+++ worker/ui/WorkerUI.hpp
@@ -279,7 +279,7 @@ public:
                 if (worker_) {
                     task_count_.store(worker_->GetTaskCount(), std::memory_order_relaxed);
                     ProcessUsage usage = worker_->GetProcessUsage();
-                    cpu_usage_.store(usage.cpu_percent, std::memory_order_relaxed);
+                    cpu_usage_.store(static_cast<float>(usage.cpu_percent), std::memory_order_relaxed);
                     mem_usage_.store(usage.memory_bytes / (1024.0f * 1024.0f), std::memory_order_relaxed);
                     auto bytes_sent = worker_->GetBytesSent();      // Already formatted (e.g., 1.5MB)
                     auto bytes_received = worker_->GetBytesReceived();
